% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut_summary.R
\name{mut_summary}
\alias{mut_summary}
\title{mut_summary}
\usage{
mut_summary(df)
}
\arguments{
\item{df}{a data.frame of mutations with mutation types}
}
\value{
a summary of counts per mutation
}
\description{
mut_summary
}
\examples{
vcffile <- system.file("extdata", "chr22.vcf.gz", package="VariantAnnotation")
vcf <- VariantAnnotation::readVcf(vcffile, "hg38")
refBase <- VariantAnnotation::ref(vcf)
refBase <- unlist(lapply(refBase, as.character))
altBase <- VariantAnnotation::alt(vcf)
altBase <- unlist(lapply(altBase, as.character))
startRef <- stats::start(vcf)
endRef <- stats::end(vcf)
chr <- GenomeInfoDb::seqnames(vcf)
Basedf <- data.frame(ref = refBase, alt=altBase, startRef= startRef, endRef= endRef, chr = chr)
Basedf$type <- apply(Basedf,1,type)
mut_summary <- mut_summary(Basedf)
mut_summary
}
