% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut_type.R
\name{mut_type}
\alias{mut_type}
\title{mut_type}
\usage{
mut_type(mutation_list, ref.genome, context_length = 3)
}
\arguments{
\item{mutation_list}{an input VCF file}

\item{ref.genome}{reference genome}

\item{context_length}{distance from mutation in ref.genome}
}
\value{
a vector of mutation signatures
}
\description{
determines for each mutation
    the corresponding mutation type in VCF file
}
\examples{
vcffile <- system.file("extdata", "chr22.vcf.gz", package="VariantAnnotation")
vcf <- VariantAnnotation::readVcf(vcffile, "hg38")
Hsapiens <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38
vcfDF <- as(MatrixGenerics::rowRanges(vcf[1:100]), "DataFrame")
muts <- mut_type(vcfDF, ref.genome = Hsapiens)
head(muts)

}
