% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutHist.R
\name{MutHist}
\alias{MutHist}
\title{MutHist}
\usage{
MutHist(counts)
}
\arguments{
\item{counts}{a data.frame of mutations counts per type}
}
\value{
a graphical output: histogram of counts per mutation type
}
\description{
plot an histogram of mutation type counts
}
\examples{
vcffile <- system.file("extdata", "chr22.vcf.gz", package="VariantAnnotation")
vcf <- VariantAnnotation::readVcf(vcffile, "hg38")
refBase <- VariantAnnotation::ref(vcf)
refBase <- unlist(lapply(refBase, as.character))
altBase <- VariantAnnotation::alt(vcf)
altBase <- unlist(lapply(altBase, as.character))
startRef <- stats::start(vcf)
endRef <- stats::end(vcf)
chr <- GenomeInfoDb::seqnames(vcf)
Basedf <- data.frame(ref = refBase, alt=altBase, startRef= startRef, endRef= endRef, chr = chr)
Basedf$type <- apply(Basedf,1,type)
mut_summary <- mut_summary(Basedf)
MutHist(mut_summary)

}
