% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{type}
\alias{type}
\title{type}
\usage{
type(df)
}
\arguments{
\item{df}{a data.frame of mutations}
}
\value{
a function to be applied on such data.frame
}
\description{
For a data.frame of mutations,provide mutation type.
 meaning:substitution, insertion or deletion
}
\examples{
vcffile <- system.file("extdata", "chr22.vcf.gz", package="VariantAnnotation")
vcf <- VariantAnnotation::readVcf(vcffile, "hg38")
refBase <- VariantAnnotation::ref(vcf)
refBase <- unlist(lapply(refBase, as.character))
altBase <- VariantAnnotation::alt(vcf)
altBase <- unlist(lapply(altBase, as.character))
startRef <- stats::start(vcf)
endRef <- stats::end(vcf)
chr <- GenomeInfoDb::seqnames(vcf)
Basedf <- data.frame(ref = refBase, alt=altBase, startRef= startRef, endRef= endRef, chr = chr)
Basedf$type <- apply(Basedf,1,type)
head(Basedf)

}
